+++ services/learn-card-network/learn-cloud-service/jest-setup.ts
@@ -1,11 +1,4 @@
 import { TextEncoder, TextDecoder } from 'util';
-import { readFile } from 'fs/promises';
-
-const mockdidkit = readFile(
-    require.resolve('@learncard/didkit-plugin/dist/didkit/didkit_wasm_bg.wasm')
-);
-
-jest.mock('./src/didkit_wasm_bg.wasm', () => mockdidkit);
 
 global.TextEncoder = TextEncoder;
 global.TextDecoder = TextDecoder;